Split actions.js into actions/{app,auth,kansa}.js
