+++ client/src/actions/app.js
@@ -1,0 +1,8 @@
+export const showMessage = (message) => ({
+  type: 'SHOW_MESSAGE',
+  message
+});
+
+export const hideMessage = () => ({
+  type: 'HIDE_MESSAGE'
+});
